+++ sdk/eventhub/azure-eventhubs/HISTORY.rst
@@ -8,7 +8,7 @@ Release History
 
 **BugFixes**
 
-- Added an option _release_partition_on_checkpoint_failure to `EPHOptions` for `EventProcessorHost` to
+- Added an option release_partition_on_checkpoint_failure to `EPHOptions` for `EventProcessorHost` to
   instruct the EventProcessorHost to fail fast on a checkpoint failure and proactively release the partition.
   This should reduce spurious reprocessing of non-checkpointed events, at the cost of a small amount of 
   additional latency if the checkpoint interruption was actually transient.